fix(test): 머지후 실패한  테스트 코드 수정

+++ src/test/java/org/plateer/fittingroomclient/cart/mapper/CartMapperTest.java
@@ -66,7 +66,7 @@ class CartMapperTest {
 
     @Test
     void deleteCartProduct() {
-        int result = cartMapper.deleteCartProduct(14L);
+        long result = cartMapper.deleteCartProduct(14L);
         log.info(result);
     }
 }
